Develop

+++ src/Services/Sale/DeleteSaleService.ts
@@ -22,7 +22,21 @@ export class DeleteSaleService{
     }
 
     async deleteItemsFromSale(id: string){
+        const sale = await this.saleRepository.FindById(id)
+
+        if(!sale){
+            throw new Error("A venda não existe")
+        }
+
+        if(!sale.open){
+            throw new Error("A venda esta fechada")
+        }
+
         const items = await this.itemRepository.ListAllFromSaleID(id)
+
+        if (items.length === 0){
+            throw new Error("A venda não possuí nenhum produto")
+        }
         
         for(let i = 0; i < items.length; i++){
             await this.itemRepository.Delete(items[i].id)
